+++ package.json
@@ -23,6 +23,7 @@
     "materialize-css": "^0.97.5",
     "mongoose": "^4.0.0",
     "mongoose-type-url": "^1.0.1",
+    "morgan": "^1.6.1",
     "passport": "^0.3.2",
     "passport-github2": "^0.1.9"
   },
