+++ server/config/credentials.js
@@ -1,0 +1,11 @@
+/**
+ * Created by VaibhavNamburi on 16/01/2016.
+ */
+
+module.exports = {
+
+    GITHUB_CLIENT_ID : "399d03e217b75d4f4625",
+    GITHUB_CLIENT_SECRET : "8e4dcd1a7c66f1db0c438aa50aa3c4af8aa4d2d5",
+    GITHUB_CALLBACK_URL: "http://localhost:3000/auth/github/callback"
+
+};
